requirements.txt
django-ckeditor
googletrans
redis
